+++ R/canvasSize.R
@@ -1,12 +1,15 @@
 ##' Get predefined canvas sizes
 ##'
 ##' @param canvas Can either be a character with one of the values
-##'     \itemize{ \item{"standard"} - A powerpoint standard size, and
-##'     two can be shown side by side - 12 by 9 inch.  \item{"wide"} -
-##'     For a single wide plot on a powerpoint slide - 16 by 9 inch
-##'     \item{"A4"} A full A4 page - 9 by 12 inch \item{"square"} As
-##'     reads - 9 by 9 inch \item{"wide-screen"} For full screen
-##'     display - 31 by 15 inch } or it can be a list with elements
+##'     \itemize{
+##' \item{"standard"} A powerpoint standard size, and
+##'     two can be shown side by side - 7.2 by 5.4 in (WxH).
+##' \item{"wide"} For a single wide plot on a powerpoint slide - 9.6 by 5.4 in
+##' \item{"A4"} A full A4 page - 5.4 by 7.2 in
+##' \item{"square"} As reads - 5.4 by 5.4 in
+##' \item{"wide-screen"} For full screen display - 18.6 by 9 in
+##' }
+##' or it can be a list with elements
 ##'     width and height with single values (unit is inches). Example:
 ##'     canvas=list(height=5,width=9).
 ##' @param scale A scale to apply to both directions of the canvas
@@ -26,11 +29,17 @@
 canvasSize <- function(canvas,scale=1,simplify=TRUE){
 
     possible.canvases <- list(
-        standard=list(width=12,height=9),
-        wide=list(width=16,height=9),
-        A4=list(width=9,height=12),
-        square=list(width=9,height=9),
-        "wide-screen"=list(width=31,height=15)
+        standard=list(width=7.2,height=5.4),
+        wide=list(width=9.6,height=5.4),
+        A4=list(width=5.4,height=7.2),
+        square=list(width=5.4,height=5.4),
+        "wide-screen"=list(width=18.6,height=9)
+
+        ## standard=list(width=12,height=9),
+        ## wide=list(width=16,height=9),
+        ## A4=list(width=9,height=12),
+        ## square=list(width=9,height=9),
+        ## "wide-screen"=list(width=31,height=15)
     )
     
     unfold.canvas <- function(canvas){