+++ R/writeFlextab.R
@@ -1,4 +1,5 @@
 #' @describeIn ftwrite Deprecated function name. Use ftwrite.
+#' @export
 
 writeFlextab <- function(ft,file,script,formats,save,quiet=FALSE,...){
     .Deprecated(new="ftwrite")